+++ src/renderer/components/settings/index.ts
@@ -1,3 +1,4 @@
 export { WalletSettings } from './WalletSettings'
 export { UnlockWalletSettings } from './UnlockWalletSettings'
 export { AppSettings } from './AppSettings'
+export { AppGeneralSettings } from './AppGeneralSettings'