+++ src/renderer/components/wallet/assets/AssetsTableCollapsable.tsx
@@ -33,6 +33,7 @@ import { isKeystoreWallet } from '../../../../shared/utils/guard'
 import { DEFAULT_WALLET_TYPE, ZERO_BASE_AMOUNT } from '../../../const'
 import { isCacaoAsset, isMayaAsset, isRuneNativeAsset, isUSDAsset } from '../../../helpers/assetHelper'
 import { getChainAsset } from '../../../helpers/chainHelper'
+import { isEvmChain } from '../../../helpers/evmHelper'
 import { getDeepestPool, getPoolPriceValue } from '../../../helpers/poolHelper'
 import { getPoolPriceValue as getPoolPriceValueM } from '../../../helpers/poolHelperMaya'
 import { hiddenString, noDataString } from '../../../helpers/stringHelper'
@@ -58,6 +59,7 @@ import { AssetIcon } from '../../uielements/assets/assetIcon'
 import { FlatButton } from '../../uielements/button'
 import { Action as ActionButtonAction, ActionButton } from '../../uielements/button/ActionButton'
 import { ReloadButton } from '../../uielements/button/ReloadButton'
+import { InfoIcon } from '../../uielements/info'
 import { QRCodeModal } from '../../uielements/qrCodeModal/QRCodeModal'
 import * as Styled from './AssetsTableCollapsable.styles'
 
@@ -618,9 +620,16 @@ export const AssetsTableCollapsable: React.FC<Props> = (props): JSX.Element => {
               </Styled.CopyLabelContainer>
             </Styled.HeaderAddress>
           </Col>
+
           <Col flex="0 1 auto" span={3} style={{ textAlign: 'right' }}>
             <Styled.HeaderLabel color={RD.isFailure(balancesRD) ? 'error' : 'gray'}>
-              {`${assetsTxt}`}
+              <span style={{ display: 'flex', alignItems: 'center', justifyContent: 'flex-end' }}>
+                {isEvmChain(chain) && (
+                  // @asgardexTeam Add Locale for tooltip
+                  <InfoIcon tooltip={'Token not showing, add contract in wallet settings'} color="primary" />
+                )}
+                <span style={{ marginLeft: isEvmChain(chain) ? '5px' : '0' }}>{assetsTxt}</span>
+              </span>
             </Styled.HeaderLabel>
           </Col>
           <Col flex="0 0 12rem" span={1}>