+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+# 1.22.6 (2024-10-6)
+
+## Updates/Fixes
+
+- fixed Maya.maya send [#378](https://github.com/asgardex/asgardex-desktop/pull/379)
+- fix Evm deriv paths [#](https://github.com/asgardex/asgardex-desktop/pull/379)
+
 # 1.22.5 (2024-09-6)
 
 ## Add
+
 - re-design application layout[#357](https://github.com/asgardex/asgardex-desktop/pull/357)
 
 ## Updates/Fixes
@@ -13,6 +21,8 @@
 - redesign asset select modal [#362](https://github.com/asgardex/asgardex-desktop/pull/362)
 - Updated electron package [electron](https://github.com/asgardex/asgardex-desktop/pull/360)
 - Added chain Radix [#294](https://github.com/asgardex/asgardex-desktop/pull/363)
+- Auto bond | node provider address checking [#348](https://github.com/asgardex/asgardex-desktop/pull/354)
+- Remove Swapping to synths, Target asset cannot be synth [#373](https://github.com/asgardex/asgardex-desktop/pull/364)
 
 # 1.22.4 (2024-09-6)